Align classic e2e specs with suite conventions
